\name{ParticleFilter-package}
\alias{ParticleFilter-package}
\alias{ParticleFilter}
\docType{package}
\title{\packageTitle{ParticleFilter}}
\description{\packageDescription{ParticleFilter}}
\section{Package Content}{\packageIndices{ParticleFilter}}
\author{\packageAuthor{ParticleFilter}}
\section{Maintainer}{\packageMaintainer{ParticleFilter}}
\keyword{package}
