% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run particle filtering models and sample trajectories of variables}
\usage{
run_model(
  theta = NULL,
  y = NULL,
  data = NULL,
  data_type = c("infection", "symptom onset", "confirmation"),
  rep = 1,
  nparticle = 100,
  tend = 2,
  dt = 0.2
)
}
\arguments{
\item{theta}{Parameter values}

\item{y}{lstate variables}

\item{data}{data to calculate likelihood against}

\item{data_type}{infection, symptom onset, or confirmation}
}
\description{
The \code{run_model()} runs particle filtering models
}
\examples{
pop <- setup_population(disease, country); calculate_cases(disease, country, pop)
}
